// src/openapi/model/mvmStockDto.ts
/**
 * Gestion de stock REST API
 * Documentation pour l\'API REST Gestion des stocks
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: wiamejaoui.wj@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProduitDto } from './produitDto';


export interface MvmStockDto { 
    id?: number;
    dateMvt?: string;
    quantite?: number;
    produit?: ProduitDto;
    typeMvt?: MvmStockDto.TypeMvtEnum;
    sourceMvt?: MvmStockDto.SourceMvtEnum;
    idEntreprise?: number;
}
export namespace MvmStockDto {
    export type TypeMvtEnum = 'ENTREE' | 'SORTIE' | 'CORRECTION_POS' | 'CORRECTION_NEG';
    export const TypeMvtEnum = {
        Entree: 'ENTREE' as TypeMvtEnum,
        Sortie: 'SORTIE' as TypeMvtEnum,
        CorrectionPos: 'CORRECTION_POS' as TypeMvtEnum,
        CorrectionNeg: 'CORRECTION_NEG' as TypeMvtEnum
    };
    export type SourceMvtEnum = 'COMMANDE_CLIENT' | 'COMMANDE_FOURNISSEUR' | 'VENTE';
    export const SourceMvtEnum = {
        CommandeClient: 'COMMANDE_CLIENT' as SourceMvtEnum,
        CommandeFournisseur: 'COMMANDE_FOURNISSEUR' as SourceMvtEnum,
        Vente: 'VENTE' as SourceMvtEnum
    };
}